Simple solution to sending traces across the network. Works for simple types it seems.
