+++ checker/src/main/kotlin/com/dickow/chortlin/checker/correlation/CorrelationValue.kt
@@ -1,3 +1,3 @@
 package com.dickow.chortlin.checker.correlation
 
-data class CorrelationValue(val key: String, val value: Any)
+data class CorrelationValue(val key: String, val value: Any?)
