+++ checker/src/main/kotlin/com/dickow/chortlin/checker/correlation/factory/CorrelationFactory.kt
@@ -16,13 +16,13 @@ object CorrelationFactory {
 
     @JvmStatic
     fun <R> correlation(method: ChortlinMethod0<*, R>, key: String, correlationFunction: CFunc0): CorrelationBuilder0<R> {
-        val func = { _: Array<Any> -> correlationFunction() }
+        val func = { _: Array<Any?> -> correlationFunction() }
         return CorrelationBuilder0(method, InputTypesFunction(key, func))
     }
 
     @JvmStatic
     fun <T1, R> correlation(method: ChortlinMethod1<*, T1, R>, key: String, correlationFunction: CFunc1<T1>): CorrelationBuilder1<T1, R> {
-        val func = { args: Array<Any> ->
+        val func = { args: Array<Any?> ->
             correlationFunction(args[0] as T1)
         }
         return CorrelationBuilder1(method, InputTypesFunction(key, func))
@@ -30,7 +30,7 @@ object CorrelationFactory {
 
     @JvmStatic
     fun <T1, T2, R> correlation(method: ChortlinMethod2<*, T1, T2, R>, key: String, correlationFunction: CFunc2<T1, T2>): CorrelationBuilder2<T1, T2, R> {
-        val func = { args: Array<Any> ->
+        val func = { args: Array<Any?> ->
             correlationFunction(args[0] as T1, args[1] as T2)
         }
         return CorrelationBuilder2(method, InputTypesFunction(key, func))
@@ -38,7 +38,7 @@ object CorrelationFactory {
 
     @JvmStatic
     fun <T1, T2, T3, R> correlation(method: ChortlinMethod3<*, T1, T2, T3, R>, key: String, correlationFunction: CFunc3<T1, T2, T3>): CorrelationBuilder3<T1, T2, T3, R> {
-        val func = { args: Array<Any> ->
+        val func = { args: Array<Any?> ->
             correlationFunction(args[0] as T1, args[1] as T2, args[2] as T3)
         }
         return CorrelationBuilder3(method, InputTypesFunction(key, func))
@@ -46,7 +46,7 @@ object CorrelationFactory {
 
     @JvmStatic
     fun <T1, T2, T3, T4, R> correlation(method: ChortlinMethod4<*, T1, T2, T3, T4, R>, key: String, correlationFunction: CFunc4<T1, T2, T3, T4>): CorrelationBuilder4<T1, T2, T3, T4, R> {
-        val func = { args: Array<Any> ->
+        val func = { args: Array<Any?> ->
             correlationFunction(args[0] as T1, args[1] as T2, args[2] as T3, args[3] as T4)
         }
         return CorrelationBuilder4(method, InputTypesFunction(key, func))
@@ -54,13 +54,13 @@ object CorrelationFactory {
 
     @JvmStatic
     fun fromInput(key: String, addFunction: CFunc0): CorrelationFunction {
-        val func = { _: Array<Any> -> addFunction() }
+        val func = { _: Array<Any?> -> addFunction() }
         return InputTypesFunction(key, func)
     }
 
     @JvmStatic
     fun <T> fromInput(key: String, addFunction: CFunc1<T>): CorrelationFunction {
-        val func = { args: Array<Any> ->
+        val func = { args: Array<Any?> ->
             addFunction(args[0] as T)
         }
         return InputTypesFunction(key, func)
@@ -68,7 +68,7 @@ object CorrelationFactory {
 
     @JvmStatic
     fun <T1, T2> fromInput(key: String, addFunction: CFunc2<T1, T2>): CorrelationFunction {
-        val func = { args: Array<Any> ->
+        val func = { args: Array<Any?> ->
             addFunction(args[0] as T1, args[1] as T2)
         }
         return InputTypesFunction(key, func)
@@ -76,7 +76,7 @@ object CorrelationFactory {
 
     @JvmStatic
     fun <T1, T2, T3> fromInput(key: String, addFunction: CFunc3<T1, T2, T3>): CorrelationFunction {
-        val func = { args: Array<Any> ->
+        val func = { args: Array<Any?> ->
             addFunction(args[0] as T1, args[1] as T2, args[2] as T3)
         }
         return InputTypesFunction(key, func)
@@ -84,7 +84,7 @@ object CorrelationFactory {
 
     @JvmStatic
     fun <T1, T2, T3, T4> fromInput(key: String, addFunction: CFunc4<T1, T2, T3, T4>): CorrelationFunction {
-        val func = { args: Array<Any> ->
+        val func = { args: Array<Any?> ->
             addFunction(args[0] as T1, args[1] as T2, args[2] as T3, args[3] as T4)
         }
         return InputTypesFunction(key, func)
@@ -92,7 +92,7 @@ object CorrelationFactory {
 
     @JvmStatic
     fun <T> fromReturn(key: String, addFunction: CFunc1<T>): CorrelationFunction {
-        val func = { returnArg: Any -> addFunction(returnArg as T) }
+        val func = { returnArg: Any? -> addFunction(returnArg as T) }
         return ReturnTypesFunction(key, func)
     }
 }
