--- core/src/main/java/module-info.java
@@ -1,11 +1,0 @@
-module chortlin.core {
-    requires transitive kotlin.stdlib;
-    requires net.bytebuddy.agent;
-    requires net.bytebuddy;
-    requires java.instrument;
-    requires kotlinx.coroutines.core;
-
-    requires transitive chortlin.shared;
-    requires transitive chortlin.checker;
-    requires transitive chortlin.interception;
-}
