--- core/src/test/kotlin/com/dickow/chortlin/core/test/shared/OnlineCheckerClasses.kt
@@ -1,16 +1,0 @@
-package com.dickow.chortlin.core.test.shared
-
-class OnlineFirstClass {
-    fun method1() {}
-    fun method2() {}
-}
-
-class OnlineSecondClass {
-    fun method1() {}
-    fun method2() {}
-}
-
-class OnlineThirdClass {
-    fun method1() {}
-    fun method2() {}
-}
