--- interception/src/main/kotlin/com/dickow/chortlin/interception/instrumentation/Instrumentation.kt
@@ -1,9 +1,0 @@
-package com.dickow.chortlin.interception.instrumentation
-
-import com.dickow.chortlin.shared.observation.ObservableParticipant
-
-
-interface Instrumentation {
-    fun before(participant: ObservableParticipant)
-    fun after(participant: ObservableParticipant)
-}
