+++ pom.xml
@@ -8,7 +8,7 @@
     <artifactId>master</artifactId>
     <version>1.0</version>
     <modules>
-        <module>core</module>
+        <module>test-application</module>
         <module>interception</module>
         <module>checker</module>
         <module>shared</module>
