+++ shared/src/main/java/module-info.java
@@ -7,4 +7,7 @@ module chortlin.shared {
     exports com.dickow.chortlin.shared.scope;
     exports com.dickow.chortlin.shared.trace;
     exports com.dickow.chortlin.shared.annotations;
+
+    exports com.dickow.chortlin.shared.trace.dto;
+    opens com.dickow.chortlin.shared.trace.dto;
 }
