+++ shared/src/main/kotlin/com/dickow/chortlin/shared/observation/Observable.kt
@@ -2,7 +2,8 @@ package com.dickow.chortlin.shared.observation
 
 import java.lang.reflect.Method
 
-abstract class Observable(val clazz: Class<*>, val method: Method) {
+abstract class Observable(public val clazz: Class<*>, val method: Method) {
+
     override fun equals(other: Any?): Boolean {
         return if (other is Observable) {
             this.clazz == other.clazz && this.method == other.method
@@ -12,7 +13,7 @@ abstract class Observable(val clazz: Class<*>, val method: Method) {
     }
 
     override fun hashCode(): Int {
-        var result = clazz.hashCode()
+        var result = method.hashCode()
         result = 31 * result + method.hashCode()
         return result
     }
