+++ shared/src/main/kotlin/com/dickow/chortlin/shared/trace/TraceElement.kt
@@ -5,5 +5,5 @@ import com.dickow.chortlin.shared.observation.Observation
 
 abstract class TraceElement {
     abstract fun getObservation(): Observation
-    abstract fun getArguments(): Array<Any>
+    abstract fun getArguments(): Array<Any?>
 }
