+++ test-application/src/main/kotlin/com/dickow/chortlin/core/CheckInMemory.kt
