+++ test-application/src/test/kotlin/com/dickow/chortlin/core/test/checker/ScopeTests.kt
